refactor(RoomCard): streamline image handling and integrate RoomDetailModal for improved code organization
